Cohesion and coupling
In OOP is a good practice work with strong cohesion and weak coupling:

Cohesion:  determines how closely related system components are in solving specific problems

-Strong is when program components access and implement logic strictly related to their purpose and interact with other components to solve a specific problem.
Implementation : defining concise classes and methods.

- is weak when its components access or implement logic that does not correspond to its nature or design purpose.

Coupling:
refers to how dependent the system components are on each other to solve a task.

-weak : is to make that a systeam has no depending between its components
Implementation :  interfaces .

-Strong : The components of the system are very dependent on each other,

For see implementation check the examples of encapsulation 05 and interfaces 04.